fix(goal): Fixed `undefined` leave messages

+++ src/Events/guildDelete.ts
@@ -8,18 +8,19 @@ import { captureException } from '@sentry/node';
 export default {
   name: 'guildDelete',
   async execute(guild: Guild) {
+    if (!guild.available) return;
     const database = getDb();
     await database.setup.deleteMany({ where: { guildId: guild.id } });
     disconnect({ serverId: guild.id });
 
-    const goalChannel = guild.client.channels.cache.get(constants.channel.goal) as TextChannel;
+    const goalChannel = guild.client.channels.cache.get(constants.channel.goal) as TextChannel | undefined;
 
-    goalChannel.send({
+    goalChannel?.send({
       embeds: [
         new EmbedBuilder()
           .setTitle('I have been kicked from a server 😢')
           .setDescription(stripIndents`
-            **${800 - guild.client.guilds.cache.size}** servers more to go! 💪
+            **${900 - guild.client.guilds.cache.size}** servers more to go! 💪
 
             **Server Name:** ${guild.name} (${guild.id})
             **Member Count:** ${guild.memberCount}
